+++ New Store/src/App.tsx
@@ -7,6 +7,7 @@ import { ProductDetails } from "./pages/ProductDetails"
 import { Navbar } from "./components/Navbar"
 import { ShoppingCartProvider } from "./context/ShoppingCartContext"
 import   './globals.css'
+import { ProductItemProps } from "./components/ProductItem";
 
 function App() {
   return (
