+++ New Store/src/components/ProductItem.tsx
@@ -5,7 +5,23 @@ import blackHeart from "../images/blackHeart.svg";
 import whiteHeart from "../images/whiteHeart.svg";
 import { useState } from "react";
 import * as C from '../styles'
-import {ProductItemProps} from '../types/product'
+import { useNavigate } from "react-router-dom";
+// import {ProductItemProps} from '../types/product'
+
+
+export type ProductItemProps = {
+  id: number;
+  title: string;
+  price: number;
+  thumbnail: string;
+  brand: string;
+  discountPercentage: number;
+  description: string;
+  rating: number;
+  stock: number;
+};
+
+
 
 export function ProductItem({
   id,
@@ -18,6 +34,8 @@ export function ProductItem({
   rating,
   stock,
 }: ProductItemProps) {
+
+  
   const {
     getItemQuantity,
     increaseCartQuantity,
@@ -29,7 +47,7 @@ export function ProductItem({
 
   const [liked, setLiked] = useState(false);
   const [show, setShow] = useState(false);
-
+  const navigate = useNavigate();
   const like = liked ? blackHeart : whiteHeart;
   const onClickCurtida = () => {
     if (liked) {
