+++ New Store/src/pages/ProductDetails.tsx
@@ -1,39 +1,14 @@
-import { Carousel , Row, Col, Card, Container, Button } from "react-bootstrap";
-import { ProductItem } from "../components/ProductItem"
-import storeItems from "../data/items.json"
-import { formatCurrency } from "../utilities/formatCurrency"
-import { CartItem } from "../components/CartItem"
-import { useShoppingCart } from "../context/ShoppingCartContext"
-import styled from 'styled-components'
+import { Carousel, Row, Col, Card, Container, Button } from "react-bootstrap";
+import { ProductItem } from "../components/ProductItem";
+import storeItems from "../data/items.json";
+import { formatCurrency } from "../utilities/formatCurrency";
+import { CartItem } from "../components/CartItem";
+import { useShoppingCart } from "../context/ShoppingCartContext";
+import styled from "styled-components";
 import { BackIcon } from "../images/icons/arrow-left-square-fill";
 import { useNavigate, useParams } from "react-router-dom";
-import {ProductItemProps} from '../types/product'
-import * as C from '../styles'
-interface ButtonProps{
-  navigate: string;
-}
-
-// const ContainerBack = styled.div`
-// display: flex;
-// justify-content: start;
-// align-items: flex-start;
-
-// `
-
-// const BackButton = styled.button`
-// display: flex;
-// justify-content: center;
-// align-items: center;
-// gap: 8px;
-// color: var(--bg-primary);
-// font-size: 14px;
-// font-weight: 500;
-
-
-// background-color: transparent;
-// border: none;
-// cursor: pointer;
-// `
+import { ProductItemProps } from "../components/ProductItem";
+import * as C from "../styles";
 
 export function ProductDetails({
   id,
@@ -51,82 +26,73 @@ export function ProductDetails({
 
   const handleBack = () => {
     navigate("/");
-};
-
+  };
 
   return (
-<>
-  
-<C.ContainerBack>
-  <C.BackButton onClick={handleBack}>
- <BackIcon />
+    <>
+      <C.ContainerBack>
+        <C.BackButton onClick={handleBack}>
+          <BackIcon />
+          Voltar
+        </C.BackButton>
+      </C.ContainerBack>
 
-    Voltar
-    </C.BackButton>
-</C.ContainerBack>
-
-
-
-    <section>
-      <img />
-     <div className="description">
-      <h5 className=""> Descrição do produto</h5>
-     </div>
-     
-    </section>
-
-    <Card className="h-100">
-      <Card.Img
-        variant="top"
-        src={thumbnail}
-        height="200px"
-        style={{ objectFit: "cover" }}
-      />
-      <Card.Body className="d-flex flex-column">
-        <Card.Title className="d-flex justify-content-between align-items-baseline mb-4">
-
-          <p className="fs-2">
-            {title}
-            </p>
-         
-          <img alt={"Icone"} 
-          // src={like} onClick={onClickCurtida} 
-          />
-          <C.Lined className="lined">
-          <p className="ms-2 text-muted ">
-            {formatCurrency(price)}
-            </p> 
-          </C.Lined>
-            <p className=""> {formatCurrency( price - ((discountPercentage *  price) / 100)  ) } </p>
-
-        </Card.Title>
-        <div className="d-flex justify-content-between">
-          <Card.Subtitle className="mb-2 ">
-            <strong>Marca:</strong>
-            {brand}
-          </Card.Subtitle>
-          <Card.Subtitle className="mb-2 ">
-            <strong>Reputação:</strong> {rating}
-          </Card.Subtitle>
+      <section>
+        <img />
+        <div className="description">
+          <h5 className=""> Descrição do produto</h5>
         </div>
-        <Card.Text className="">
-          <p className="fs-2">
-            {description}
-            </p>
-        </Card.Text>
-        <Card.Subtitle className="mb-2 ">
-          {stock}
-          <strong> em estoque</strong>
-        </Card.Subtitle>
+      </section>
 
-        <div className="mt-auto">
-          {/* {quantity === 0 ? ( */}
-            <Button className="w-100" 
-            // onClick={() => increaseCartQuantity(id)}
+      <Card className="h-100">
+        <Card.Img
+          variant="top"
+          src={thumbnail}
+          height="200px"
+          style={{ objectFit: "cover" }}
+        />
+        <Card.Body className="d-flex flex-column">
+          <Card.Title className="d-flex justify-content-between align-items-baseline mb-4">
+            <p className="fs-2">{title}</p>
+
+            <img
+              alt={"Icone"}
+              // src={like} onClick={onClickCurtida}
+            />
+            <C.Lined className="lined">
+              <p className="ms-2 text-muted ">{formatCurrency(price)}</p>
+            </C.Lined>
+            <p className="">
+           
+              {formatCurrency(price - (discountPercentage * price) / 100)}{" "}
+            </p>
+          </Card.Title>
+          <div className="d-flex justify-content-between">
+            <Card.Subtitle className="mb-2 ">
+              <strong>Marca:</strong>
+              {brand}
+            </Card.Subtitle>
+            <Card.Subtitle className="mb-2 ">
+              <strong>Reputação:</strong> {rating}
+            </Card.Subtitle>
+          </div>
+          <Card.Text className="">
+            <p className="fs-2">{description}</p>
+          </Card.Text>
+          <Card.Subtitle className="mb-2 ">
+            {stock}
+            <strong> em estoque</strong>
+          </Card.Subtitle>
+
+          <div className="mt-auto">
+            {/* {quantity === 0 ? ( */}
+            <Button
+              className="w-100"
+              // onClick={() => increaseCartQuantity(id)}
             >
               + Adicionar no Carinho
             </Button>
-          {/* ) : ( */}
+            {/* ) : ( */}
             <div
               className="d-flex align-items-center flex-column"
               style={{ gap: ".5rem" }}
@@ -135,18 +101,23 @@ export function ProductDetails({
                 className="d-flex align-items-center justify-content-center"
                 style={{ gap: ".5rem" }}
               >
-                <Button 
+                <Button
                 // onClick={() => decreaseCartQuantity(id)}
-                >-</Button>
+                >
+                  -
+                </Button>
                 <div>
                   <p className="fs-3">
                     {/* {quantity}  */}
-                    qtd 
-                    </p> no carrinho
+                    qtd
+                  </p>{" "}
+                  no carrinho
                 </div>
-                <Button 
+                <Button
                 // onClick={() => increaseCartQuantity(id)}
-                >+</Button>
+                >
+                  +
+                </Button>
               </div>
               <Button
                 // onClick={() => removeFromCart(id)}
@@ -156,22 +127,10 @@ export function ProductDetails({
                 Remover
               </Button>
             </div>
-          {/* )} */}
-        </div>
-      </Card.Body>
-    </Card>
-
-  {/* {cartItems.map(item => (
-    <CartItem key={item.id} {...item} />
-    ))}
-     
-            {formatCurrency(
-              cartItems.reduce((total, cartItem) => {
-                const item = storeItems.find(i => i.id === cartItem.id)
-                return total + (item?.price || 0) * cartItem.quantity
-              }, 0)
-              )} */}
- </>
-
-  )
+            {/* )} */}
+          </div>
+        </Card.Body>
+      </Card>
+    </>
+  );
 }
