+++ New Store/src/types/product.tsx
@@ -1,16 +1,16 @@
 
 
-  export type ProductItemProps = {
-    id: number;
-    title: string;
-    price: number;
-    thumbnail: string;
-    brand: string;
-    discountPercentage: number;
-    description: string;
-    rating: number;
-    stock: number;
-  };
+  // export type ProductItemProps = {
+  //   id: number;
+  //   title: string;
+  //   price: number;
+  //   thumbnail: string;
+  //   brand: string;
+  //   discountPercentage: number;
+  //   description: string;
+  //   rating: number;
+  //   stock: number;
+  // };
